根据时间范围查询报表

+++ xc-trade/src/main/java/com/xc/trade/entity/vo/ReportsByRangeVO.java
@@ -26,7 +26,7 @@ public class ReportsByRangeVO {
     /**
      * 成交额
      */
-    private BigDecimal tradedAmount;
+    private Integer tradedAmount;
     /**
      * 成交数量
      */
